Split out credentialed database tests
